Finishing MeasurementScreen refactor with new architecture
