+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/MeasurementScreenViewModel.kt
@@ -1,0 +1,13 @@
+package org.noiseplanet.noisecapture.ui.features.measurement
+
+import androidx.lifecycle.ViewModel
+import org.noiseplanet.noisecapture.measurements.MeasurementsService
+
+class MeasurementScreenViewModel(
+    private val measurementsService: MeasurementsService,
+) : ViewModel() {
+
+    fun startRecordingAudio() = measurementsService.startRecordingAudio()
+
+    fun stopRecordingAudio() = measurementsService.stopRecordingAudio()
+}
