+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/plot/PlotAxisBuilder.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.ui.features.measurement
+package org.noiseplanet.noisecapture.ui.features.measurement.plot
 
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextMeasurer
